Fix a mountain of 64bit windows warnings
The console version now starts though there are still a pile of pointer truncation issues that are sure to render making connections impossible and operation in general quite unstable. p.dll and the other console tools have not been ported yet.

Changes tested against Visual C++ 1.0, 2.0, 6.0, 2002, 2022, OpenWatcom 1.9

+++ kermit/k95/ckoco2.c
@@ -5252,7 +5252,7 @@ Win32ConsoleInit( void )
         return;
 
     AllocConsole();
-    hCrt = _open_osfhandle((long) GetStdHandle(STD_OUTPUT_HANDLE),
+    hCrt = _open_osfhandle((intptr_t) GetStdHandle(STD_OUTPUT_HANDLE),
                                 _O_BINARY);
     hf = _fdopen( hCrt, "wb" );
     *stdout = *hf;
@@ -5266,12 +5266,12 @@ Win32ConsoleInit( void )
                              OPEN_EXISTING, FILE_ATTRIBUTE_NORMAL, NULL ) ;
         SetStdHandle( STD_INPUT_HANDLE, handle ) ;
     }
-    hCrt = _open_osfhandle((long) handle, _O_BINARY);
+    hCrt = _open_osfhandle((intptr_t ) handle, _O_BINARY);
     hf = _fdopen( hCrt, "rb" );
     *stdin = *hf;
     i = setvbuf( stdin, NULL, _IONBF, 0 );
 
-    hCrt = _open_osfhandle((long) GetStdHandle(STD_ERROR_HANDLE), _O_BINARY);
+    hCrt = _open_osfhandle((intptr_t ) GetStdHandle(STD_ERROR_HANDLE), _O_BINARY);
     hf = _fdopen( hCrt, "wb" );
     *stderr = *hf;
     i = setvbuf( stderr, NULL, _IONBF, 0 );
